pyproject.toml
[project]
name = "strands-agents-example"
version = "0.1.0"
description = "Strands Agents Example"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Paola Arce"}
]
dependencies = [
    "click",
    "strands-agents>=0.1.2",
    "strands-agents-tools>=0.1.1",
]

[dependency-groups]
dev = [
    "black",
    "ipykernel",
    "ipywidgets",
    "isort",
    "pre-commit",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
naming_agent = "strands_agents_example.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strands_agents_example"]

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.black]
line-length = 79
preview = true
skip-string-normalization = false
include = '\.pyi?$'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
profile = "black"
line_length = 79
skip = ["__init__.py"]
